Improved fake plugin loading process.
