+++ Tests/PHPCI/Plugin/Util/Fake/ExamplePluginWithNoConstructorArgs.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Tests\PHPCI\Plugin\Util\Fixtures;
+
+use PHPCI\Plugin;
+
+class ExamplePluginWithNoConstructorArgs implements Plugin
+{
+    public function execute()
+    {
+    }
+}
