+++ Tests/PHPCI/Plugin/Util/Fake/ExamplePluginWithSingleOptionalArg.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Tests\PHPCI\Plugin\Util\Fixtures;
+
+use PHPCI\Plugin;
+
+class ExamplePluginWithSingleOptionalArg implements Plugin
+{
+    function __construct($optional = null)
+    {
+
+    }
+
+    public function execute()
+    {
+
+    }
+}
