+++ Tests/PHPCI/Plugin/Util/Fake/ExamplePluginWithSingleTypedRequiredArg.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Tests\PHPCI\Plugin\Util\Fixtures;
+
+use PHPCI\Plugin;
+
+class ExamplePluginWithSingleTypedRequiredArg implements Plugin
+{
+
+    public $RequiredArgument;
+
+    function __construct(\stdClass $requiredArgument)
+    {
+        $this->RequiredArgument = $requiredArgument;
+    }
+
+    public function execute()
+    {
+
+    }
+}
